Replaced static DBFacade with singleton DBFacade. Prepared for case of DB absence. This can be expanded further upon.
