+++ src/Chirp.Razor/DBFacade.cs
@@ -1,64 +1,74 @@
 using System.Globalization;
-
 using Microsoft.Data.Sqlite;
-using SQLitePCL;
 
-namespace Chirp.Razor
+namespace Chirp.Razor;
+
+class DBFacade : ICheepService, IDisposable
 {
-    public class DBFacade
+    private const int CHEEPS_PER_PAGE = 32;
+    private readonly SqliteConnection Connection;
+
+    public DBFacade()
     {
+        string sqlDBFilePath = Environment.GetEnvironmentVariable("CHIRPDBPATH") ?? CreateTemporaryDatabase();
+        Connection = new SqliteConnection($"Data Source={sqlDBFilePath}");
+        Connection.Open();
+    }
+    
+    public List<CheepViewModel> GetCheeps(int page)
+    {
+        List<CheepViewModel> cheeps = new List<CheepViewModel>();
+        var command = Connection.CreateCommand();
+        command.CommandText = "SELECT u.username, m.text, m.pub_date FROM user u, message m WHERE u.user_id = m.author_id ORDER by m.pub_date desc LIMIT " + CHEEPS_PER_PAGE + " OFFSET " + getPageOffset(page);
+
+        using var reader = command.ExecuteReader();
+        while (reader.Read())
+        {
+            Object[] values = new Object[reader.FieldCount];
+            int fieldCount = reader.GetValues(values);
+            for (int i = 0; i < fieldCount; i += 3)
+                cheeps.Add(new CheepViewModel($"{values[i]}", $"{values[i+1]}", $"{UnixTimeStampToDateTimeString(Convert.ToInt64(values[i+2]))}"));
+        }
+        return cheeps;
+    }
+
+    public List<CheepViewModel> GetCheepsFromAuthor(string author, int page)
+    {
+        List<CheepViewModel> cheeps = new List<CheepViewModel>();
+        var command = Connection.CreateCommand();
+        command.CommandText = $"SELECT u.username, m.text, m.pub_date FROM user u, message m WHERE u.user_id = m.author_id AND u.username = '{author}' ORDER by m.pub_date desc LIMIT " + CHEEPS_PER_PAGE + " OFFSET " + getPageOffset(page);
         
-        private static int getoffset(int page)
+        using var reader = command.ExecuteReader();
+        while (reader.Read())
         {
-            return (page - 1) * 32;
+            Object[] values = new Object[reader.FieldCount];
+            int fieldCount = reader.GetValues(values);
+            for (int i = 0; i < fieldCount; i += 3)
+                cheeps.Add(new CheepViewModel($"{values[i]}", $"{values[i+1]}", $"{UnixTimeStampToDateTimeString(Convert.ToInt64(values[i+2]))}"));
         }
-        private static List<CheepViewModel> getcheeps(string sqlQuery)
-        {
-            List<CheepViewModel> cheeps = new List<CheepViewModel>();
-            // SQLite provider needs to be initialized before making any calls to SQLite
-            // Call to SQLitePCL.Batteries.Init() initializes the necessary provider for SQLite before attempting to open a connection to the database.
-            // Information retrieved from ChatGPT.
-            Batteries.Init();
+        return cheeps;
+    }
 
-            string sqlDBFilePath = Environment.GetEnvironmentVariable("CHIRPDBPATH") ?? throw new NullReferenceException("No ChirpDBPATH environment variable set");
-            
-            using var connection = new SqliteConnection($"Data Source={sqlDBFilePath}");
-            {
-                connection.Open();
+    private string CreateTemporaryDatabase()
+    {
+        return Path.GetTempPath() + "chirp.db";
+    }
 
-                var command = connection.CreateCommand();
-                command.CommandText = sqlQuery;
+    public void Dispose()
+    {
+        Connection.Dispose();
+    }
 
-                using var reader = command.ExecuteReader();
-                while (reader.Read())
-                {
-                    Object[] values = new Object[reader.FieldCount];
-                    int fieldCount = reader.GetValues(values);
-                    for (int i = 0; i < fieldCount; i += 3)
-                        cheeps.Add(new CheepViewModel($"{values[i]}", $"{values[i+1]}", $"{UnixTimeStampToDateTimeString(Convert.ToInt64(values[i+2]))}"));
-                } 
-            }
-            return cheeps;
-        }
-        public static List<CheepViewModel> Read(int page)
-        {
-            int offset = getoffset(page);
-            return getcheeps(@"SELECT u.username, m.text, m.pub_date FROM user u, message m WHERE u.user_id = m.author_id ORDER by m.pub_date desc LIMIT 32 OFFSET " + offset);
-        }
-        public static List<CheepViewModel> UserRead(string author, int page)
-        {
-            author = $"'{author}'";
-            int offset = getoffset(page);
-            return getcheeps(
-                $"SELECT u.username, m.text, m.pub_date FROM user u, message m WHERE u.user_id = m.author_id AND u.username = {author} ORDER by m.pub_date desc LIMIT 32 OFFSET " + offset);
-        }
-        
-        public static string UnixTimeStampToDateTimeString(long unixTimeStamp)
-        {
-            DateTimeOffset dateTimeOffset = DateTimeOffset.FromUnixTimeSeconds(unixTimeStamp);
-            TimeZoneInfo danishTimeZone = TimeZoneInfo.FindSystemTimeZoneById("Central Europe Standard Time");
-            DateTime dateTime = TimeZoneInfo.ConvertTime(dateTimeOffset.UtcDateTime, danishTimeZone);
-            return dateTime.ToString("dd/MM/yyyy HH:mm:ss", CultureInfo.InvariantCulture);
-        }
+    public static string UnixTimeStampToDateTimeString(long unixTimeStamp)
+    {
+        DateTimeOffset dateTimeOffset = DateTimeOffset.FromUnixTimeSeconds(unixTimeStamp);
+        TimeZoneInfo danishTimeZone = TimeZoneInfo.FindSystemTimeZoneById("Central Europe Standard Time");
+        DateTime dateTime = TimeZoneInfo.ConvertTime(dateTimeOffset.UtcDateTime, danishTimeZone);
+        return dateTime.ToString("dd/MM/yyyy HH:mm:ss", CultureInfo.InvariantCulture);
+    }
+
+    private static int getPageOffset(int page)
+    {
+        return (page - 1) * CHEEPS_PER_PAGE;
     }
 }
